cluster/master.py
"""
Master Node - Coordinates the entire cluster

Responsibilities:
- Accept tasks from users/APIs
- Analyze task requirements
- Route tasks to appropriate worker nodes
- Aggregate results
- Maintain cluster health
"""
import asyncio
import structlog
from typing import Optional, Dict, Any
from pydantic import BaseModel

logger = structlog.get_logger()


class TaskRequest(BaseModel):
    """Incoming task from user"""
    task_id: str
    description: str
    priority: int = 3  # 1=highest, 5=lowest
    required_capabilities: list[str] = []
    context: Dict[str, Any] = {}


class MasterNode:
    """
    Master node that coordinates distributed agent cluster
    """
    
    def __init__(self, config: Dict[str, Any]):
        self.config = config
        self.node_id = config.get("node_id", "master-001")
        self.workers: Dict[str, Dict] = {}  # worker_id -> metadata
        self.task_queue: asyncio.Queue = asyncio.Queue()
        logger.info("master_node_initialized", node_id=self.node_id)
    
    async def start(self):
        """Start master node services"""
        logger.info("starting_master_node")
        
        # Start service discovery
        await self._start_discovery()
        
        # Start gRPC server
        await self._start_grpc_server()
        
        # Start task dispatcher
        await self._start_task_dispatcher()
        
        logger.info("master_node_ready")
    
    async def _start_discovery(self):
        """Initialize service discovery (Redis/etcd)"""
        logger.info("starting_service_discovery")
        # TODO: Implement Redis-based discovery
        pass
    
    async def _start_grpc_server(self):
        """Start gRPC server for worker connections"""
        logger.info("starting_grpc_server", port=self.config.get("grpc_port", 50051))
        # TODO: Implement gRPC server
        pass
    
    async def _start_task_dispatcher(self):
        """Start task dispatcher loop"""
        logger.info("starting_task_dispatcher")
        # TODO: Implement task dispatcher
        pass
    
    async def submit_task(self, task: TaskRequest) -> str:
        """
        Submit a task to the cluster
        
        Returns:
            task_id for tracking
        """
        logger.info("task_submitted", task_id=task.task_id)
        await self.task_queue.put(task)
        return task.task_id
    
    async def get_task_result(self, task_id: str) -> Optional[Dict[str, Any]]:
        """Get result of a completed task"""
        # TODO: Implement result retrieval
        pass
    
    def register_worker(self, worker_id: str, capabilities: list[str], address: str):
        """Register a new worker node"""
        self.workers[worker_id] = {
            "capabilities": capabilities,
            "address": address,
            "status": "active",
            "tasks_running": 0
        }
        logger.info("worker_registered", worker_id=worker_id, capabilities=capabilities)
    
    def select_worker(self, required_capabilities: list[str]) -> Optional[str]:
        """
        Select best worker for a task based on capabilities and load
        
        Returns:
            worker_id or None if no suitable worker found
        """
        # TODO: Implement intelligent worker selection
        # Consider: capabilities match, current load, latency, etc
        pass


async def main():
    """Entry point for master node"""
    import yaml
    
    with open("config/master.yaml") as f:
        config = yaml.safe_load(f)
    
    master = MasterNode(config)
    await master.start()
    
    # Keep running
    try:
        await asyncio.Event().wait()
    except KeyboardInterrupt:
        logger.info("master_node_shutting_down")


if __name__ == "__main__":
    asyncio.run(main())
