requirements.txt
grpcio>=1.60.0
grpcio-tools>=1.60.0
redis>=5.0.1
psycopg2-binary>=2.9.9
pgvector>=0.2.4
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
PyYAML>=6.0.1
structlog>=24.1.0
typer>=0.9.0
rich>=13.7.0

# AI & LLM
openai>=1.10.0
anthropic>=0.18.0
litellm>=1.30.0

# Optional: Hermes Agent base (if we want to inherit functionality)
# git+https://github.com/NousResearch/hermes-agent.git

# Development
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
black>=23.12.0
ruff>=0.1.0
mypy>=1.7.0
